requirements_mac.txt
absl-py==2.0.0
adagio==0.2.4
adtk==0.6.2
ansi2html==1.8.0
antlr4-python3-runtime==4.11.1
anyio==4.1.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
catboost==1.2.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmdstanpy==1.2.0
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.6
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.46.0
fqdn==1.5.1
fs==2.4.16
fsspec==2023.12.2
fugue==0.8.7
fugue-sql-antlr==0.2.0
gast==0.5.4
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.60.0
h5py==3.10.0
holidays==0.38
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-dash==0.4.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
lightgbm==4.1.0
lightning-utilities==0.10.1
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
MetPy==1.6.1
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nfoursid==1.0.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.4
pexpect==4.9.0
Pillow==10.1.0
Pint==0.23
platformdirs==4.1.0
plotly==5.18.0
pmdarima==2.0.4
pooch==1.8.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
prophet==1.1.5
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyod==1.1.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qpd==0.4.4
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
Send2Trash==1.8.2
shap==0.44.1
six==1.16.0
slicer==0.0.7
sniffio==1.3.0
soupsieve==2.5
sqlglot==20.10.0
stack-data==0.6.3
stanio==0.3.0
statsforecast==1.7.1
statsmodels==0.14.0
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tbats==1.1.3
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
torch==2.1.2
torchmetrics==1.3.0.post0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
Tree==0.2.4
triad==0.9.4
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
u8darts==0.27.2
uri-template==1.3.0
urllib3==2.1.0
utilsforecast==0.0.26
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xarray==2023.12.0
xgboost==2.0.3
zipp==3.17.0
